// types/ProductForm.ts
export type FormDataType = {
  name: string;
  price: Number;
  sku: string;
  weight: string;
  stock: Number;
  description: string;
  salePrice: string;
  genericName: string;
  packSize: string;
  strength: string;
  dosageForm: string;
  manufacturer: string;
  prescriptionReq: string;
  status: string;
  expiryDate: string;
  images: FileList;
};